nuke /lib dir - move all logic into index.js
